exposed epochs and added read-function.
Epochs are exposed in the `Reader`-class via property `epochs`.  A
method `get_physical_samples_from_epoch` is added that receives an
instance of `xml_files.Epoch` and returns all data in it.  Optional
arguments `t0`, `dt`, and `channels` can be provided.  As opposed to
`get_physical_samples`, `t0` is relative to `Epoch.t0`.