Fixed Epoch handling by read function.
`xml_files.Epoch` contains time/duration information and block indices.
The block indices are now used to access the data, whereas the time
information is, informational with regards to a discontinuous .mff file.

Furthermore, we fixed the RawBinFile data reading test.  Issue was that
`A` was not reindexed correctly as `np.searchsorted`-kwarg `side` was
changed from `left` to `right`.