Refactored props with cache.
The `cached_property` pattern is used to store caches of files and
intermediate results, which were explicitely cached previously.

The reduction in code size increases maintainability, but keep
in mind that no explicite recaching is triggered whhen computations
in the code change.  For this one needs to `del` the cache.

The cache of a property `p` is implicitely stored in the object as an
attribute `_cached_p`.