Do not fix indexing in raw_bin_files
Results from `get_physical_samples` and
`get_physical_samples_from_epoch` cannot be compared.  The former is an
internal API call that should not be used.  It is accessed by the
letter, which we support in the official API.

When accessing only `get_physical_samples`, (a,b) need to be reindexed
when going across block boundaries.

`a, b` (`t0, dt`) are relative to the epoch that one wants to access.
API `get_physical_samples` disrespects that.  The returned values for
the real `start` do not reflect the temporal offset that can happen in
between consecutive epochs.

Other changes
-------------

- Specify version of numpy==1.15.1
- Encorporate `pytest.approx` where possible