Reindex a, b in raw_bin_file.
This commit fixes a serious bug in `RawBinFile.read_raw_samples`.

It is an amendement to the reasoning given in the previous commit with
hash `334f73b14e7e7a4faed103bd3d570093313ad2fa`, which was faulty.

Explanantion
------------

Parameter `block_slice` is determined by the continous epoch that is
read out, and `t0,dt` determine the relative time wrt the beginning of
that epoch.  First, temporal indices `a,b` are computed wrt the start of
the epoch.  Then, the blocks `A,B` are computed that enclose `a,b`.  In
the last step, the read-out blocks were referenced by `a,b`, but
remember that `a,b` were relative to the epoch start, not these blocks.

To fix the wrong indexing, we now subtract from `a,b` the offset created
by only reading blocks `A,B`, which is `bsi[A]-bsi[0]` (bsi:
`block_start_idx`).