Read in multiple .bin files
This commit is a minor change to the
`_blobs` function in the `Reader` class.
Before, when we looped through the .bin
files all .bin files defaulted to 'EEG'
for signal_type. We now get the signal_type
from the info.xml file that corresponds to
the .bin file in question.