Add `name` attribute to `Epoch` class
Here we add an attribute `Epoch.name` to the
`mffpy.epoch.Epochs` class. This is initialized
to be "epoch" for any epoch.

When reading in an MFF file, `Epoch.name` is
automatically populated with the category name
corresponding to the epoch (if this information
is available) when running `Reader.epochs`.

We add a method to the `Reader` class called
`epochs_by_name` which returns a list of all
epochs matching a given name. We also add a
private helper function `_sort_categories_by_starttime`
to sort the category info by start time, allowing
us to easily match this info up to the epochs.