Add tests for epochs and categories
Here we expand the test cases for
`mffpy/tests/test_reader.py` to test
the functionality of `Reader.categories`,
`Reader.epochs`, and `Reader.epochs_by_name`.

We add a fourth example MFF file that is
an averaged file so that we can test
retrieving epochs info from both
segmented and averaged MFF files.