Clean up and documentation
Here we get rid of a couple of unnecessary
functions in `mffpy.reader` that have been
moved to `mffpy.xml_files`.

We add a function `Categories.sort_categories_by_starttime`
to give us a list of sorted blocks with category
names. We use this function when associating
categories and epochs.

We also add some docstrings and tests.