Make reading and writing congruent
Here we make it so that a `Categories.categories`
dict that is read from an XML file can fit
perfectly into `Categories.content` to be written
out.

We add 'name' as an element of a segment within
a category and make it so that optional segment
elements are not read in if they are None.